requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
Cython==0.29.16
graphviz==0.8.4
idna==2.9
imutils==0.5.3
multidict==4.7.5
mxnet-cu102==1.6.0
numpy==1.18.2
opencv-python==4.2.0.34
python-engineio==3.12.1
python-socketio==4.5.1
requests==2.23.0
six==1.14.0
termcolor==1.1.0
urllib3==1.25.8
yarl==1.4.2
